requirements.txt
aiohttp==3.5.3
arrow==0.12.1
atomicwrites==1.2.1
attrs==18.2.0
Automat==0.7.0
bitstruct==6.0.0
certitude==1.0.1
Click==7.0
constantly==15.1.0
EasyProcess==0.2.3
git+https://github.com/ebroecker/canmatrix@8f63047c18dcb65f7904d298cbf603d2e6325b85#egg=canmatrix
future==0.17.1
gitdb2==2.0.5
GitPython==2.1.11
git+https://github.com/altendky/graham@89f3b3e6aad3b98bafa83f28fbbc11bb92df6ad8#egg=graham
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
lxml==4.2.5
marshmallow==2.16.3
more-itertools==4.3.0
natsort==5.5.0
Pint==0.8.1
pluggy==0.8.0
py==1.7.0
pyelftools==0.25
PyHamcrest==1.9.0
PyQt5==5.11.3
PyQt5-sip==4.19.13
pytest==4.0.1
pytest-qt==3.2.1
pytest-rerunfailures==5.0
pytest-xvfb==1.1.0
git+https://github.com/altendky/pytest-twisted@6f04ba3ad806b3b40602999e71c28633c36d81f9#egg=pytest-twisted
python-can==3.0.0
python-dateutil==2.7.5
python-docx==0.8.7
python-dotenv==0.9.1
PyVirtualDisplay==0.2.1
git+https://github.com/altendky/qt5reactor@ce3ea2dad6d71adb10cde79b64bc0c7d27e9139c#egg=qt5reactor
six==1.11.0
smmap2==2.0.5
treq==18.6.0
Twisted==18.9.0
typing==3.6.6
wrapt==1.10.11
zope.interface==4.6.0

-e .
